feat: Finish detail screen
