+++ lib/src/constants/typography.dart
@@ -89,3 +89,38 @@ Text poppinsText(
     ),
   );
 }
+
+class FarmSwapSectionTitle extends StatelessWidget {
+  const FarmSwapSectionTitle({
+    super.key,
+    required this.height,
+    required this.title,
+    this.onPress,
+  });
+
+  final double height;
+  final String title;
+  final VoidCallback? onPress;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        poppinsText(
+          value: title,
+          size: 15 / 812 * height,
+          isBold: true,
+        ),
+        const Spacer(),
+        GestureDetector(
+          onTap: onPress,
+          child: poppinsText(
+            value: 'See All',
+            size: 15 / 812 * height,
+            color: const Color(0xFF14BE77),
+          ),
+        ),
+      ],
+    );
+  }
+}
