+++ lib/src/features/dashboard/presentation/dashboard_screen.dart
@@ -1,11 +1,13 @@
 import 'package:farmswap_v2/src/constants/colors.dart';
 import 'package:farmswap_v2/src/features/dashboard/presentation/bottom_nav.dart';
+import 'package:farmswap_v2/src/features/listing/presentation/item_detail_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../../common_widgets/farm_swap_buttons/farmswap_white_icon.dart';
 import '../../../common_widgets/input/farmswap_searchbar.dart';
 import '../../../constants/typography.dart';
+import '../../listing/presentation/list_item_card.dart';
+import '../../listing/presentation/list_item_horizontal_cart.dart';
 
 class DasboardScreen extends StatelessWidget {
   const DasboardScreen({super.key});
@@ -230,6 +232,15 @@ class DasboardScreen extends StatelessWidget {
                               productRating: "4.5",
                               productDistance: "1.2",
                               totalReviews: "20",
+                              onTap: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                    builder: (context) =>
+                                        const ItemDetailScreen(),
+                                  ),
+                                );
+                              },
                             ),
                             const SizedBox(
                               width: 20,
@@ -500,229 +511,6 @@ class CircleCategory extends StatelessWidget {
   }
 }
 
-class HorizontalListItem extends StatelessWidget {
-  HorizontalListItem({
-    super.key,
-    required this.width,
-    required this.height,
-    required this.productName,
-    required this.productImage,
-    required this.productPrice,
-    required this.productRating,
-    required this.productDistance,
-    required this.totalReviews,
-    this.promoLabel = "FOR SWAP ONLY",
-  });
-
-  final double width;
-  final double height;
-  final String productName;
-  final String productImage;
-  final String productPrice;
-  final String productRating;
-  final String productDistance;
-  final String totalReviews;
-  String promoLabel;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 320 / 375 * width,
-      padding: const EdgeInsets.all(6),
-      constraints: const BoxConstraints(
-        maxWidth: 400,
-        maxHeight: 115,
-        minWidth: 320,
-        minHeight: 65,
-      ),
-      decoration: ShapeDecoration(
-        color: Colors.white,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        shadows: const [
-          BoxShadow(
-            color: Color(0x11000000),
-            blurRadius: 50,
-            offset: Offset(26, 26),
-            spreadRadius: 0,
-          )
-        ],
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: [
-          Container(
-            width: 124 / 375 * width,
-            height: 100 / 812 * height,
-            constraints: const BoxConstraints(
-              maxWidth: 150,
-              minWidth: 94,
-              minHeight: 65,
-            ),
-            child: Stack(
-              children: [
-                Positioned.fill(
-                  child: Container(
-                    decoration: BoxDecoration(
-                      image: DecorationImage(
-                        image: AssetImage(
-                          productImage,
-                        ),
-                        fit: BoxFit.cover,
-                      ),
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                  ),
-                ),
-                Positioned(
-                  top: 5,
-                  left: 10,
-                  child: Container(
-                    // width: 28,
-                    // height: 10,
-                    padding: const EdgeInsets.symmetric(
-                      vertical: 2,
-                      horizontal: 5,
-                    ),
-                    decoration: ShapeDecoration(
-                      color: const Color(0xFF14BE77),
-                      shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(4)),
-                      shadows: const [
-                        BoxShadow(
-                          color: Color(0x3FFFFFFF),
-                          blurRadius: 40,
-                          offset: Offset(15, 20),
-                          spreadRadius: 0,
-                        )
-                      ],
-                    ),
-                    child: poppinsText(
-                      value: promoLabel,
-                      size: 8 / 812 * height,
-                      color: Colors.white,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          const SizedBox(
-            width: 20,
-          ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: [
-              farmSwapFont(
-                text: "8 hours : 12 minutes",
-                size: 18 / 812 * height,
-                color: FarmSwapGreen.normalGreen,
-              ),
-              SizedBox(
-                height: 10 / 812 * height,
-              ),
-              SizedBox(
-                child: poppinsText(
-                  value: productName,
-                  size: 12 / 812 * height,
-                ),
-              ),
-              SizedBox(
-                height: 10 / 812 * height,
-              ),
-              poppinsText(
-                value: '$productDistance km | ⭐ $productRating ($totalReviews)',
-                size: 10 / 812 * height,
-                color: const Color(0xFF5F5D6B),
-              ),
-              SizedBox(
-                height: 10 / 812 * height,
-              ),
-              Row(
-                children: [
-                  poppinsText(
-                    value: productPrice,
-                    size: 14 / 812 * height,
-                    isBold: true,
-                    color: FarmSwapGreen.normalGreen,
-                  ),
-                  const SizedBox(
-                    width: 3,
-                  ),
-                  poppinsText(
-                    value: "|",
-                    size: 10 / 812 * height,
-                    color: const Color(0xFF5F5D6B),
-                  ),
-                  const SizedBox(
-                    width: 3,
-                  ),
-                  SizedBox(
-                    child: Align(
-                      child: SvgPicture.asset(
-                        height: 15,
-                        width: 15,
-                        "assets/svg/dashboard/delivery icon.svg",
-                        fit: BoxFit.cover,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(
-                    width: 3,
-                  ),
-                  poppinsText(
-                    value: "₱20.00",
-                    size: 10 / 812 * height,
-                    color: const Color(0xFF5F5D6B),
-                  ),
-                ],
-              )
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class FarmSwapSectionTitle extends StatelessWidget {
-  const FarmSwapSectionTitle({
-    super.key,
-    required this.height,
-    required this.title,
-    this.onPress,
-  });
-
-  final double height;
-  final String title;
-  final VoidCallback? onPress;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        poppinsText(
-          value: title,
-          size: 15 / 812 * height,
-          isBold: true,
-        ),
-        const Spacer(),
-        GestureDetector(
-          onTap: onPress,
-          child: poppinsText(
-            value: 'See All',
-            size: 15 / 812 * height,
-            color: const Color(0xFF14BE77),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
 class ScrollableRecommendedBadges extends StatelessWidget {
   const ScrollableRecommendedBadges({
     super.key,
@@ -850,172 +638,6 @@ class RecommendBadge extends StatelessWidget {
 }
 
 // ignore: must_be_immutable
-class ListItemCard extends StatelessWidget {
-  ListItemCard({
-    super.key,
-    required this.productName,
-    required this.productImage,
-    required this.productPrice,
-    required this.productRating,
-    required this.productDistance,
-    required this.totalReviews,
-    this.promoLabel = "PROMO",
-    required this.height,
-    required this.width,
-  });
-
-  final String productName;
-  final String productImage;
-  final String productPrice;
-  final String productRating;
-  final String productDistance;
-  final String totalReviews;
-  String promoLabel;
-  final double height;
-  final double width;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(6),
-      width: 150 / 375 * width,
-      constraints: const BoxConstraints(
-        maxWidth: 200,
-        minWidth: 150,
-        minHeight: 200,
-      ),
-      decoration: ShapeDecoration(
-        color: Colors.white,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        shadows: const [
-          BoxShadow(
-            color: Color(0x11000000),
-            blurRadius: 50,
-            offset: Offset(0, 26),
-            spreadRadius: 0,
-          )
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          SizedBox(
-            height: 120 / 812 * height,
-            child: Stack(
-              children: [
-                Positioned.fill(
-                  child: Container(
-                    decoration: BoxDecoration(
-                      image: DecorationImage(
-                        image: AssetImage(
-                          productImage,
-                        ),
-                        fit: BoxFit.cover,
-                      ),
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                  ),
-                ),
-                Positioned(
-                  top: 5,
-                  left: 10,
-                  child: Container(
-                    // width: 28,
-                    // height: 10,
-                    padding: const EdgeInsets.symmetric(
-                      vertical: 2,
-                      horizontal: 5,
-                    ),
-                    decoration: ShapeDecoration(
-                      color: const Color(0xFF14BE77),
-                      shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(4)),
-                      shadows: const [
-                        BoxShadow(
-                          color: Color(0x3FFFFFFF),
-                          blurRadius: 40,
-                          offset: Offset(15, 20),
-                          spreadRadius: 0,
-                        )
-                      ],
-                    ),
-                    child: poppinsText(
-                      value: promoLabel,
-                      size: 8 / 812 * height,
-                      color: Colors.white,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          SizedBox(
-            height: 10 / 812 * height,
-          ),
-          SizedBox(
-            width: 96,
-            child: poppinsText(
-              value: productName,
-              size: 12 / 812 * height,
-            ),
-          ),
-          SizedBox(
-            height: 10 / 812 * height,
-          ),
-          poppinsText(
-            value: '$productDistance km | ⭐ $productRating ($totalReviews)',
-            size: 10 / 812 * height,
-            color: const Color(0xFF5F5D6B),
-          ),
-          SizedBox(
-            height: 10 / 812 * height,
-          ),
-          Row(
-            children: [
-              poppinsText(
-                value: productPrice,
-                size: 14 / 812 * height,
-                isBold: true,
-                color: FarmSwapGreen.normalGreen,
-              ),
-              const SizedBox(
-                width: 3,
-              ),
-              poppinsText(
-                value: "|",
-                size: 10 / 812 * height,
-                color: const Color(0xFF5F5D6B),
-              ),
-              const SizedBox(
-                width: 3,
-              ),
-              SizedBox(
-                child: Align(
-                  child: SvgPicture.asset(
-                    height: 15,
-                    width: 15,
-                    "assets/svg/dashboard/delivery icon.svg",
-                    fit: BoxFit.cover,
-                  ),
-                ),
-              ),
-              const SizedBox(
-                width: 3,
-              ),
-              poppinsText(
-                value: "₱20.00",
-                size: 10 / 812 * height,
-                color: const Color(0xFF5F5D6B),
-              ),
-            ],
-          )
-        ],
-      ),
-    );
-  }
-}
 
 class CategoryIcon extends StatelessWidget {
   const CategoryIcon({
